fix: try fetch lost layer when unpack image
If someone removes the content or snapshot by-pass the cri plugin, cir in-memory metadata think
the image exists. When we create sandbox or container, we can't find snapshot and unpack from the
content layer with the `discard_unpacked_layers` cri config. In that case, we can try fetch the lost
image layer to complete create container.

Signed-off-by: Yadong Ding <ding_yadong@foxmail.com>